added thumbnail images and updated to nextjs app router organization
