+++ app/blog/page.tsx
@@ -1,7 +1,7 @@
-import BlogCard from "../components/BlogCard/BlogCard";
-import Intro from "../components/Intro/Intro";
+import BlogCard from "../../components/BlogCard/BlogCard";
+import Intro from "../../components/Intro/Intro";
 import styles from "./page.module.css";
-import { getAllPosts } from "../lib/utils";
+import { getAllPosts } from "../../lib/utils";
 // import { BlogPostProps } from "../lib/types";
 
 export default async function Home() {
