--- app/components/BlogCard/BlogCard.module.css
@@ -1,17 +1,0 @@
-.container {
-  width: 100%;
-}
-
-.title{
-  font-size: xx-large;
-}
-
-.information{
-  color: grey;
-  font-size: large;
-}
-
-.tags{
-  flex: 1;
-  background-color: lightgray;
-}
