--- app/components/BlogLayout/BlogLayout.tsx
@@ -1,7 +1,0 @@
-import React, { ReactNode } from "react";
-
-const BlogLayout = ({ children }: { children: ReactNode }) => {
-  return <div className="blog-layout">{children}</div>;
-};
-
-export default BlogLayout;
