--- app/components/Intro/Intro.module.css
@@ -1,6 +1,0 @@
-.Intro {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 1em 0 0;
-}
