--- app/components/Intro/Intro.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import styles from './Intro.module.css'
-
-const Intro = () => {
-  return (
-    <div className={styles.Intro}>Welcome to my blog! I write about all things beauty, clothing, and my life as a student. </div>
-  )
-}
-
-export default Intro
