--- app/posts/glossiers-welcome-to-sephora.md
@@ -1,17 +1,0 @@
----
-title: Glossier's Welcome to Sephora
-categories: [beauty]
-publishedDate: 02/18/2023
-excerpt: Glossier is coming to Sephora stores on February 23rd
-tags: beauty, glossier, sephora
----
-
-Glossier’s arrival in Canada has been long awaited. I know all other Canadian beauty lovers have been waiting for this day since 2017 when they launched a small pop-up in Toronto. Finally, fans can get their hands on the skincare-first makeup-second brand on February 23rd.
-
-Glossier has been a direct-to-consumer brand from day one. Although they’ve opened pop-up and permanent retail locations in the states, the brand's identity has lived on the internet for international customers.
-
-Not only will Glossier be available in Canada, but products will be easily accessible at Sephora locations. More benefits include not needing to pay for shipping, the ability to test products in person, and taking advantage of beauty insider rewards.
-
-While I’ve only tried a select number of products from their line, my personal favourites are their [cloud paint](https://www.glossier.com/products/cloud-paint) and [balm dotcom](https://www.glossier.com/products/balm-dotcom).
-
-I hope you’re as excited as I am to see that iconic millennial pink in stores.
