--- app/posts/sample.md
@@ -1,9 +1,0 @@
----
-title: Sample Post
-categories: [misc]
-publishedDate: 12/26/2024
-excerpt: some excerpt words
-tags: test, blog, react
----
-
-Welcome to my awesome blog. Trying to get this [markdown thing](https://bionicjulia.com/blog/setting-up-nextjs-markdown-blog-with-typescript) to work.
