--- app/posts/whats-in-my-bag.md
@@ -1,53 +1,0 @@
----
-title: What's in My Bag
-categories: [favourites]
-publishedDate: 04/02/2023
-excerpt: Some insight into a few of my favourites!
-tags: bag, beauty
----
-
-Bags are custom to every individual, that it’s so intimate finding what others keep in theirs. I love discovering everyone’s favorite items that go beyond the beauty world. Here I’m sharing what I keep in mine.
-
-My bag is the black shoulder bag from [Brandy Melville](https://us.brandymelville.com/products/mini-faux-leather-purse). I find it’s the perfect bag that goes with most any outfit.
-
-## Essentials
-
-_[Wallet](https://www.clutchhealdsburg.com/products/lodis-card-holder)_
-
-No explanation necessary.
-
-_Hand Sanitizer_
-
-After COVID, this is non-negotiable.
-
-_Phone_
-
-Also, no explanation necessary.
-
-_Hair Tie_
-
-An easy hairstyle in a pinch during a windy day.
-
-## Beauty
-
-_[Kosas Lip Oil- Jellyfish](https://www.sephora.com/ca/en/product/lip-oil-P451920?skuId=2297505&icid2=products%20grid:p451920:product)_
-
-Dry lips are my biggest pet peeve-- if I happen to forget my lip oil then I’m scouring for a huge bottle of water.
-
-_[It Cosmetics Bye Bye Undereye Concealer- 20.0 Medium](https://www.sephora.com/ca/en/product/bye-bye-under-eye-full-coverage-anti-aging-waterproof-concealer-mini-P446901?skuId=2224053&icid2=products%20grid:p446901:product)_
-
-You never know when you need a touch up, and this mini size fits in my bag perfectly.
-
-## Miscellaneous
-
-_Pocket Mirror_
-
-To make sure I look decent when a bathroom isn’t in my proximity.
-
-_[Pen](https://muji.ca/collections/pens-pencils/products/wooden-mechanical-pen)_
-
-Always handy.
-
-_[Airpods](https://www.apple.com/ca/shop/product/MQD83AM/A/airpods-pro?fnode=d2ddc7a306450c651694de68e41b0f1c1dffccb745f654d8d535650ff7f7c71a457bbb562dde5e23986a0a11e4fc0bea25adbe2cdd27b81d8dadefd6a4e1fa3669f99b628632e4f0db42987cb5c33ba2fe1652d961dcd34b2ceda4963c9a00d3b2ad6fd170e1390ee67425157153953341deb2d2ca6a095562544bd7a60e196f&fs=f%3Dapple-overear-sport%26fh%3D47d1%252B3214%252B45aa%252B45ab)_
-
-Waiting in line, avoiding droning small talk, you know, the usual.
