+++ components/BlogCard/BlogCard.module.css
@@ -1,0 +1,38 @@
+.container {
+  width: 100%;
+  display: flex;
+  justify-content: left;
+  align-items: flex-start;
+  min-height: 300px;
+  margin: 1em 0;
+}
+
+.imageWrapper{
+  position: relative;
+  width: 400px; 
+  height: 300px;
+  display: flex;
+  margin: 0 1em;
+}
+
+.image{
+  object-fit: cover;
+}
+
+.text{
+  padding: 1em;
+}
+
+.title{
+  font-size: xx-large;
+}
+
+.information{
+  color: grey;
+  font-size: large;
+}
+
+.tags{
+  flex: 1;
+  background-color: lightgray;
+}
