+++ components/Footer/Footer.module.css
@@ -1,0 +1,33 @@
+.wrapper {
+    background-color: #1d2024;
+}
+
+.Footer {
+    color: #c3c3c3;
+    display: flex;
+    width: 80%;
+    justify-content: space-between;
+    padding: 2em;
+    margin: 8em auto 0;
+}
+
+.Footer ul{
+    display: flex;  
+    width: 100%;  
+}
+
+.Footer li {
+    flex: 1;
+    text-align: left;
+}
+
+.Footer h1 {
+    color: #fefefe;
+    font-size: 1.5em;
+}
+
+.subList {
+    display: flex;
+    flex-direction: column;
+    line-height: 2.5em;
+}
