+++ public/posts/overrated-makeup-products.md
@@ -1,0 +1,20 @@
+---
+title: Overrated Makeup Products
+excerpt: Makeup that is just not worth the hype.
+categories: [beauty]
+publishedDate: 07/04/2023
+tags: beauty, kosas, lip-gloss, makeup, makeup-forever, overrated, powder, saie
+thumbnail: "/images/overrated-makeup.jpg"
+---
+
+**Saie Airset Radiant Loose Setting Powder**
+
+This product looks flawless initially. But after a few minutes, it makes your makeup look dry, cakey, and orange. I have the lightest shade but it never fails to make me look like a pumpkin.
+
+**Kosas Wet Lip Oil Treatment Plumping Gloss**
+
+Honestly an amazing gloss, but the product leaks absolutely everywhere. I even featured it in my "What's In My Bag", but it made a sticky mess in my purse after a few days. Great product, terrible packaging.
+
+**Make Up For Ever HD Skin All-In-One Palette**
+
+I was so excited to try this palette after seeing Hailey Bieber use the same one. But the formula lacks creaminess. It applies patchy and doesn't look natural at all.
