+++ public/posts/quintessential-summer-lip-colours.md
@@ -1,0 +1,28 @@
+---
+title: Quintessential Summer Lip Colours
+excerpt: Juicy summer lip product picks.]
+categories: [beauty]
+publishedDate: 07/06/2023
+tags: beauty, glossier, laneige, lip, lip-balm, lip-gloss, makeup, marc-jacobs, r-e-m-beauty, summer, tower-28
+thumbnail: "/images/summer-lip.jpg"
+---
+
+**Lip Sleeping Mask (Berry)**
+
+This lip mask moisturizes your lips while looking so deliciously glossy. It completely represents the epitome of summer makeup. I'm obsessed with the berry scent for the summer.
+
+**Glossier Balm Dotcom (Mango)**
+
+This is also a lip balm like the Laneige mask, but a slightly less decadent version. I don't find this formula particularly hydrating, but the delicious mango flavour makes up for it. The packaging is adorable and easy for on-the-go. The shade supposedly has a peachy-rose tint, but I just find it to be clear. 
+
+**Tower 28 JuiceBalm (Shake)**
+
+I love using this tinted lip balm when I want some colour. "Shake" is a juicy sheer pink. I also wouldn't reach for this if my lips were seriously dry. But I love it for its sheer tint and convenience.
+
+**r.e.m. beauty Lip Liner Pencil (Key Change)**
+
+During the day, I wear this layered with the Laneige lip mask to soften it. But if I'm going out for the night then it's nice to have a colour more potent. This shade is a dusty rose and is a great MLBB.
+
+**Marc Jacobs Enamoured Lip Lacquer (Pretty Thing)**
+
+This product is easily my favourite lip gloss by far. It's a pigmented, nude-rose shade with amazing shine. But it kills me because I'm pretty sure it's discontinued. When I first used it, I was expecting a warm, vanilla scent, but it's actually a refreshing minty smell.
