--- src/data/isten-to-the-best-songs-for-april-2023.md
@@ -1,21 +1,0 @@
----
-title: Listen to the Best Songs for April 2023
-excerpt: Give a listen to some of my new favourite music for spring.
-category: favourites
-publishedDate: 04/18/2023
-tags:
-  - april
-  - favourites
-  - music
-  - playlist
-  - spring
----
-
-Some of my newly discovered to long-time favourite songs I’m listening to this month.
-
-- [Count to Five- Bedroom](https://open.spotify.com/track/4ynqq6nbZUPI0fsD5ooALh?si=553a7ed09819453a)
-- [Selfless- The Strokes](https://open.spotify.com/track/2t0wwvR15fc3K1ey8OiOaN?si=97a1bb60ccda4f5c)
-- [Silver Soul- Beach House](https://open.spotify.com/track/2LZ2CNbcbJrwh53pitRgom?si=b44c3179869a4458)
-- [Fall in Love with You.- Montell Fish](https://open.spotify.com/track/4kJT7Yj6Za01KfKHjb7mZE?si=6f807d1c5f8f41c4)
-- [Only She Knows- Loving](https://open.spotify.com/track/2AXd4r3Y00PjjNxceN3mZm?si=eaeac8aa7e2f4517)
-- [Witches- Alice Phoebe Lou](https://open.spotify.com/track/5BBK1WoaiD3b3sJ2UysLuM?si=2d81836a3a954125)
