--- src/data/prom-2023.md
@@ -1,33 +1,0 @@
----
-title: Prom 2023
-excerpt: Prom products and details!
-category: beauty, clothing
-publishedDate: 07/04/2023
-tags:
-  - grwm
-  - makeup
-  - prom
-  - shoes
----
-
-**Outfit**
-
-Of all of high school, I was most excited about prom. I pined dress sites for months leading up to it and I made mood boards of inspiration. My biggest goal was to find a look I loved but didn't cost as much. I chose a sage green dress with a corset and slit from Windsor. I was actually hung over the decision of an amazing dress from Boutique 1861. But I chose Windsor because of the unique style and price point. I got my shoes last minute from Oak + Fort. My dress ran a little long so I knew I wanted to have platform shoes. I opted for the Platform Open Toe Heel in White.
-
-**Hair**
-
-I decided on curling my hair with a titanium barrel wand to get loose, light curls. I considered some more comprehensive styles, but the light fluffiness was my favourite.
-
-**Makeup**
-
-*Face*
-
-I used a damp beauty blender to apply light amounts of the Rare Beauty Positive Light Tinted Moisturizer in 24N. This gave me a nice base, to begin with. Then I used the Kosas Revealer Super Creamy + Bright Concealer in 2.5C. For blush, I used the tarte SEA Breezy Cream Blush in Pink Sky. To add some radiance, I used the Milk Makeup Cream Highlighter Stick in Lit. To finish off the face, I lightly dusted the e.l.f. Halo Glow Setting Powder in Medium. 
-
-*Eyes*
-
-I curled the lashes with the Shiseido Eyelash Curler and coated with the Glossier Lash Slick in Black.
-
-*Lips*
-
-Finished off with the Laneige Lip Sleeping Mask in Berry.
