--- src/data/tiramisu.md
@@ -1,30 +1,0 @@
----
-title: Tiramisu
-excerpt: Easiest tiramisu recipe straight from Veneto, Italy.
-category: cooking
-publishedDate: 04/07/2024
-tags:
-  - baking
-  - dessert
-  tiramisu
----
-
-I recently made this tiramisu recipe and I can't find it anywhere so I'm sharing it here. Tiramisu seems like a complicated dessert, but it's actually very simple. It's made up of layers of ladyfingers, coffee, and a cream. If you want to go the full mile, then make your own ladyfingers. But otherwise, impress all your friends and family with this easy tiramisu.
-
-*Ingredients*
-
-- 2 Packs Lady Fingers
-- 3 Eggs
-- 60g Granulated Sugar
-- 300g Mascarpone Cheese
-- 1 Cup Strong Coffee 
-- 2 Tablespoons Liqueur (of your choice)
-- Cocoa Powder
-
-*Steps*
-1. Separate the whites and yolks of the eggs.
-2. Cream together the egg yolks and sugar.
-3. Mix in the mascarpone cheese with the egg and sugar mixture.
-4. Mix together the coffee and liqueur.
-5. Layer the ladyfingers flat in a medium sized dish. Dribble the coffee mixture onto each ladyfinger. Spread the mascarpone mixture on top. Repeat until the dish is filled. Usually about 2-3 layers.
-6. Finish off the cake with a generous dusting of cocoa powder.
